Created traits for common shared functionality
