+++ lib/govuk_design_system_formbuilder.rb
@@ -1,8 +1,14 @@
+require 'govuk_design_system_formbuilder/traits/collection_item'
+require 'govuk_design_system_formbuilder/traits/conditional'
+require 'govuk_design_system_formbuilder/traits/error'
+require 'govuk_design_system_formbuilder/traits/hint'
+require 'govuk_design_system_formbuilder/traits/label'
+require 'govuk_design_system_formbuilder/traits/localisation'
+require 'govuk_design_system_formbuilder/traits/supplemental'
+
 require 'govuk_design_system_formbuilder/version'
 require 'govuk_design_system_formbuilder/builder'
 require 'govuk_design_system_formbuilder/base'
-
-require 'govuk_design_system_formbuilder/traits/collection_item'
 
 require 'govuk_design_system_formbuilder/elements/hint'
 require 'govuk_design_system_formbuilder/elements/label'
