+++ lib/govuk_design_system_formbuilder/base.rb
@@ -1,28 +1,14 @@
 module GOVUKDesignSystemFormBuilder
   class Base
+    include Traits::Localisation
+
     delegate :capture, :content_tag, :safe_join, :tag, :raw, :link_to, to: :@builder
 
     def initialize(builder, object_name, attribute_name, &block)
-      @builder = builder
-      @object_name = object_name
+      @builder        = builder
+      @object_name    = object_name
       @attribute_name = attribute_name
-      @block_content = capture { block.call } if block_given?
-    end
-
-    def hint_element
-      @hint_element ||= Elements::Hint.new(@builder, @object_name, @attribute_name, @hint_text)
-    end
-
-    def error_element
-      @error_element ||= Elements::ErrorMessage.new(@builder, @object_name, @attribute_name)
-    end
-
-    def label_element
-      @label_element ||= Elements::Label.new(@builder, @object_name, @attribute_name, **@label)
-    end
-
-    def supplemental_content
-      @supplemental_content ||= Containers::Supplemental.new(@builder, @object_name, @attribute_name, @block_content)
+      @block_content  = capture { block.call } if block_given?
     end
 
     # returns the id value used for the input
@@ -46,45 +32,9 @@ module GOVUKDesignSystemFormBuilder
       end
     end
 
-    def hint_id
-      return nil if @hint_text.blank?
-
-      build_id('hint')
-    end
-
-    def error_id
-      return nil unless has_errors?
-
-      build_id('error')
-    end
-
-    def conditional_id
-      build_id('conditional')
-    end
-
-    def supplemental_id
-      return nil if @block_content.blank?
-
-      build_id('supplemental')
-    end
-
-    # Provides an id for use by the textual description of character and word limits.
-    #
-    # @note In order for the GOV.UK Frontend JavaScript to pick up this associated field
-    #   it has to have the same id as the text area with the additional suffix of '-info'
-    def limit_id
-      [field_id(link_errors: true), 'info'].join('-')
-    end
-
     def has_errors?
       @builder.object.errors.any? &&
         @builder.object.errors.messages.dig(@attribute_name).present?
-    end
-
-    def wrap_conditional(block)
-      content_tag('div', class: conditional_classes, id: conditional_id) do
-        capture { block.call }
-      end
     end
 
     def described_by(*ids)
@@ -92,20 +42,6 @@ module GOVUKDesignSystemFormBuilder
     end
 
   private
-
-    def localised_text(context)
-      key = localisation_key(context)
-
-      return nil unless I18n.exists?(key)
-
-      I18n.translate(key)
-    end
-
-    def localisation_key(context)
-      return nil unless @object_name.present? && @attribute_name.present?
-
-      ['helpers', context, @object_name, @attribute_name].join('.')
-    end
 
     # Builds the values used for HTML id attributes throughout the builder
     #
