+++ lib/govuk_design_system_formbuilder/elements/text_area.rb
@@ -43,6 +43,14 @@ module GOVUKDesignSystemFormBuilder
         end
       end
 
+      # Provides an id for use by the textual description of character and word limits.
+      #
+      # @note In order for the GOV.UK Frontend JavaScript to pick up this associated field
+      #   it has to have the same id as the text area with the additional suffix of '-info'
+      def limit_id
+        [field_id(link_errors: true), 'info'].join('-')
+      end
+
       def limit?
         @max_words || @max_chars
       end
