+++ lib/govuk_design_system_formbuilder/traits/collection_item.rb
@@ -1,16 +1,14 @@
 module GOVUKDesignSystemFormBuilder
-  module Elements
-    module Traits
-      module CollectionItem
-      private
+  module Traits
+    module CollectionItem
+    private
 
-        def retrieve(item, method)
-          case method
-          when Symbol, String
-            item.send(method)
-          when Proc
-            method.call(item)
-          end
+      def retrieve(item, method)
+        case method
+        when Symbol, String
+          item.send(method)
+        when Proc
+          method.call(item)
         end
       end
     end
