+++ lib/govuk_design_system_formbuilder/traits/conditional.rb
@@ -1,0 +1,15 @@
+module GOVUKDesignSystemFormBuilder
+  module Traits
+    module Conditional
+      def conditional_id
+        build_id('conditional')
+      end
+
+      def wrap_conditional(block)
+        content_tag('div', class: conditional_classes, id: conditional_id) do
+          capture { block.call }
+        end
+      end
+    end
+  end
+end
