+++ lib/govuk_design_system_formbuilder/traits/supplemental.rb
@@ -1,0 +1,15 @@
+module GOVUKDesignSystemFormBuilder
+  module Traits
+    module Supplemental
+      def supplemental_content
+        @supplemental_content ||= Containers::Supplemental.new(@builder, @object_name, @attribute_name, @block_content)
+      end
+
+      def supplemental_id
+        return nil if @block_content.blank?
+
+        build_id('supplemental')
+      end
+    end
+  end
+end
